Version 1.0-exp3 (#9)
* Refactoring - moving DSLin its own package, removing empty tests

* Main scope moved to separate file, implements IrcStateProvider

+ MainScope now does not connect Tmi automatically

* Container, TwitchPlugin interface, ReconnectPlugin

Container - Scope responsible for containing all inside it and control all the 
  TwitchMessage traffic. It's modifiable with TwitchPlugins which can add plug & play 
  functionality
TwitchPlugin - Pluggable object that transforms or filters incoming and/or outgoing 
  messages

* Docs link in README, Reconnect plugin fix

Reconnect plugin listeners fix

Documentation link added to README

Ktor 1.2.2

* Rolling back to Ktor 1.2.1

* TwitchIRC JVM and JS implementations, no Ktor, TwitchIRC changes

TwitchIRC interface without suspend functions now
JVM and JS TwitchIRC implementations
Got rid of Ktor dependencies
JVM - OkHttp dependency
JS - 'ws' npm dependency

Reconnect plugin update - clean active channels after reconnect
Reconnect plugin minor changes, IRC impl - disconnect before connecting


Reconnect delete print

* Merge from master

* Mutex in FlowDispenser for handling listOfCallbacks

* Synchronized access to callback list

* ContextScopes, ChannelScope helpers deleted, MessageContext in listeners

GlobalContextScope created

ContextScopes now filter according to their funtionality

Command helpers in ChannelScope were deleted in favour of MessageContext helpers

ChannelContext and UserContext used as receivers in message listeners and offer command helper functions for specific message received instead of scope

* UserNotice listeners, MessageContext changes, optimized imports

* Small fixes

* broadcaster, moderators, subscribers scopes

Broadcater scope was changed so it does not rely on ChannelScope
'Moderators' and 'Subscribers' scopes introduced

* Merge from master 0.0.8

* reduced duplicate code, whisper helper command

Events and HelperScopes reworked to reduce duplicate code by creating internal functions with logic and Scope specific functions rely on this function

* Experimental builds, refactoring

* Refactoring, readme changes, documentation, gradle bintray changes

* Gradle publish script fix

* isMod, isSub, isBroad... ext func to UserState, simplified helper scopes

* Kolin 1.3.50, CommandScope, MessageContext helpers - 0.1.0-exp1

* Throttle plugin and Throttle CommandScope filter - 0.1.0-exp2

* Changing the emote id from Int to String - 0.1.0-exp3

* Readme version change

* Updated data type from Int to String (#8)

+++ src/commonMain/kotlin/com/ktmi/tmi/messages/Utility.kt
@@ -49,7 +49,7 @@ fun String.parseTwitchPairSet() = this
     .split(",")
     .map { it.split("/") }
     .filter { it.size == 2 }
-    .map { it[0] to it[1].toInt() }
+    .map { it[0] to it[1] }
     .toMap()
 
 /**
